Try to make MovingPlatformHandler work vertically.
My efforts here include: displacing the raycast with the character's velocity (hmm, this isn't good to have for going down--I think it still works down, though?); laying groundwork to have a more proper velocity-tracking-based system, which will probably have the pleasant side effect of reducing GetComponent calls per-frame (on average, anyway--still gotta use it some but).

Gonna lay that out real quick: on detect platform with velocity, that platform's Velocity is stored, then _unless a new platform is detected_ that's _confirmed intercepting_ (ooh, should allow configure intercept-allowed or not! But that'd be best if it _only affects semi-solids_), or the character is _confirmed_ no longer on the platform (which should throw an event so we can have X-velocity follow into the air movement if you'd like it to be that way), strong movement should just duplicate the movement of the platform.